LangChain/competitor-research-tool/requirements.txt
aiofiles==24.1.0
aiohttp==3.11.11
altair==5.5.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.98
botocore>=1.35.98
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.4.21
cryptography==44.0.0
dataclasses-json==0.6.7
dateparser==1.2.0
defusedxml==0.7.1
fastapi==0.115.6
filelock==3.16.1
fireworks-ai==0.15.11
frozenlist==1.5.0
fsspec==2024.12.0
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-aws==0.2.10
langchain-community==0.3.14
langchain-core==0.3.29
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
litellm==1.58.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
multidict==6.1.0
nltk==3.9.1
numpy==2.2.1
openai==1.59.7
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==10.4.0
playwright==1.49.1
pluggy==1.5.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests>=2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
selenium==4.1.0
semantic-version==2.10.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
streamlit>=1.24.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3>=1.25.4,<3
uvicorn==0.34.0
webdriver-manager==3.5.0
websockets==14.1
wheel==0.44.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
